Fix breakage caused by proper uint64 implementation
Before this was implemented, uint64 was basically a alias for
int64, which meant that it *could* return negative values.

The tests depend on that behaviour, and the tests would die
because now one can *not* store a negative value in a uint64
anymore.

Fixed by just using an int64 in signatures instead.